Fixed bug #16774: t3ver_stage should be int(11) instead of tinyint(4)
git-svn-id: https://svn.typo3.org/TYPO3v4/Core/trunk@9895 709f56b5-9817-0410-a4d7-c38de5d9e867

+++ t3lib/stddb/tables.sql
@@ -157,7 +157,7 @@ CREATE TABLE pages (
   t3ver_wsid int(11) DEFAULT '0' NOT NULL,
   t3ver_label varchar(255) DEFAULT '' NOT NULL,
   t3ver_state tinyint(4) DEFAULT '0' NOT NULL,
-  t3ver_stage tinyint(4) DEFAULT '0' NOT NULL,
+  t3ver_stage int(11) DEFAULT '0' NOT NULL,
   t3ver_count int(11) DEFAULT '0' NOT NULL,
   t3ver_tstamp int(11) DEFAULT '0' NOT NULL,
   t3ver_swapmode tinyint(4) DEFAULT '0' NOT NULL,
